+++ GeneralFunctions.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <string>
 
-std::string StringConvert(int value);
+std::string StringConvert(float value);
 
 float BitConvert64(float input);
